Http client cpr

See merge request oai/cn5g/oai-cn5g-udr!64

+++ src/api_server/impl/AMF3GPPAccessRegistrationDocumentApiImpl.cpp
@@ -69,7 +69,7 @@ void AMF3GPPAccessRegistrationDocumentApiImpl::create_amf_context3gpp(
     Pistache::Http::ResponseWriter& response) {
   nlohmann::json response_data = {};
   Pistache::Http::Code code    = {};
-  long http_code               = 0;
+  uint32_t http_code           = 0;
 
   m_udr_app->handle_create_amf_context_3gpp(
       ueId, amf3GppAccessRegistration, response_data, http_code);
@@ -88,7 +88,7 @@ void AMF3GPPAccessRegistrationDocumentApiImpl::query_amf_context3gpp(
     Pistache::Http::ResponseWriter& response) {
   nlohmann::json response_data = {};
   Pistache::Http::Code code    = {};
-  long http_code               = 0;
+  uint32_t http_code           = 0;
 
   m_udr_app->handle_query_amf_context_3gpp(ueId, response_data, http_code);
