+++ src/api_server/impl/AccessAndMobilitySubscriptionDataDocumentApiImpl.cpp
@@ -59,7 +59,7 @@ void AccessAndMobilitySubscriptionDataDocumentApiImpl::query_am_data(
     Pistache::Http::ResponseWriter& response) {
   nlohmann::json response_data = {};
   Pistache::Http::Code code    = {};
-  long http_code               = 0;
+  uint32_t http_code           = 0;
 
   m_udr_app->handle_query_am_data(
       ueId, servingPlmnId, response_data, http_code);
