+++ src/api_server/impl/AuthenticationStatusDocumentApiImpl.cpp
@@ -52,7 +52,7 @@ void AuthenticationStatusDocumentApiImpl::create_authentication_status(
     Pistache::Http::ResponseWriter& response) {
   nlohmann::json response_data = {};
   Pistache::Http::Code code    = {};
-  long http_code               = 0;
+  uint32_t http_code           = 0;
 
   m_udr_app->handle_create_authentication_status(
       ueId, authEvent, response_data, http_code);
@@ -69,7 +69,7 @@ void AuthenticationStatusDocumentApiImpl::delete_authentication_status(
     const std::string& ueId, Pistache::Http::ResponseWriter& response) {
   nlohmann::json response_data = {};
   Pistache::Http::Code code    = {};
-  long http_code               = 0;
+  uint32_t http_code           = 0;
 
   m_udr_app->handle_delete_authentication_status(
       ueId, response_data, http_code);
@@ -88,7 +88,7 @@ void AuthenticationStatusDocumentApiImpl::query_authentication_status(
     Pistache::Http::ResponseWriter& response) {
   nlohmann::json response_data = {};
   Pistache::Http::Code code    = {};
-  long http_code               = 0;
+  uint32_t http_code           = 0;
 
   m_udr_app->handle_query_authentication_status(ueId, response_data, http_code);
