+++ src/api_server/impl/AuthenticationSubscriptionDocumentApiImpl.cpp
@@ -59,7 +59,7 @@ void AuthenticationSubscriptionDocumentApiImpl::
         Pistache::Http::ResponseWriter& response) {
   nlohmann::json response_data = {};
   Pistache::Http::Code code    = {};
-  long http_code               = 0;
+  uint32_t http_code           = 0;
 
   m_udr_app->handle_modify_authentication_subscription(
       ueId, patchItem, response_data, http_code);
@@ -79,7 +79,7 @@ void AuthenticationSubscriptionDocumentApiImpl::
         Pistache::Http::ResponseWriter& response) {
   nlohmann::json response_data = {};
   Pistache::Http::Code code    = {};
-  long http_code               = 0;
+  uint32_t http_code           = 0;
 
   m_udr_app->handle_read_authentication_subscription(
       ueId, response_data, http_code);
