+++ src/api_server/impl/SMFRegistrationDocumentApiImpl.cpp
@@ -53,7 +53,7 @@ void SMFRegistrationDocumentApiImpl::create_smf_context_non3gpp(
     Pistache::Http::ResponseWriter& response) {
   nlohmann::json response_data = {};
   Pistache::Http::Code code    = {};
-  long http_code               = 0;
+  uint32_t http_code           = 0;
 
   m_udr_app->handle_create_smf_context_non_3gpp(
       ueId, pduSessionId, smfRegistration, response_data, http_code);
@@ -67,7 +67,7 @@ void SMFRegistrationDocumentApiImpl::delete_smf_context(
     Pistache::Http::ResponseWriter& response) {
   nlohmann::json response_data = {};
   Pistache::Http::Code code    = {};
-  long http_code               = 0;
+  uint32_t http_code           = 0;
 
   m_udr_app->handle_delete_smf_context(
       ueId, pduSessionId, response_data, http_code);
@@ -86,7 +86,7 @@ void SMFRegistrationDocumentApiImpl::query_smf_registration(
     Pistache::Http::ResponseWriter& response) {
   nlohmann::json response_data = {};
   Pistache::Http::Code code    = {};
-  long http_code               = 0;
+  uint32_t http_code           = 0;
 
   m_udr_app->handle_query_smf_registration(
       ueId, pduSessionId, response_data, http_code);
