+++ src/api_server/impl/SMFRegistrationsCollectionApiImpl.cpp
@@ -53,7 +53,7 @@ void SMFRegistrationsCollectionApiImpl::query_smf_reg_list(
     Pistache::Http::ResponseWriter& response) {
   nlohmann::json response_data = {};
   Pistache::Http::Code code    = {};
-  long http_code               = 0;
+  uint32_t http_code           = 0;
 
   m_udr_app->handle_query_smf_reg_list(ueId, response_data, http_code);
